Refactor utility code to be shared.
