+++ tests/test_max_age.py
@@ -8,17 +8,7 @@ import pytest
 from requests import Session
 from cachecontrol.adapter import CacheControlAdapter
 from cachecontrol.cache import DictCache
-
-
-class NullSerializer(object):
-
-    def dumps(self, request, response, body=None):
-        return response
-
-    def loads(self, request, data, body_file=None):
-        if data and getattr(data, "chunked", False):
-            data.chunked = False
-        return data
+from .utils import NullSerializer
 
 
 class TestMaxAge(object):
